Remove extractor.
Ascension is now a comparator generator that always requires an array
argument. Extraction has been moved to a module called `whittle` on NPM.
I found that I'd be extracting values in different ways for different
inputs, that is that a single comparator would require multiple
extraction strategies, so we've separated comparision from extraction.

Closes #44.